QDoc: Use QVector instead of QList
Replace the use of QVector in most of QDoc.
Also, remove one redundant C-style cast to int for result from method
call that returns an int. As this happened in a macro, the result is
removing a whole bunch of nagging from code inspection.

Fixes: QTBUG-80669

+++ src/qdoc/cppcodemarker.cpp
@@ -495,7 +495,7 @@ QString CppCodeMarker::addMarkUp(const QString &in,
             keywords.insert(keywordTable[j]);
     }
 #define readChar() \
-    ch = (i < (int)code.length()) ? code[i++].cell() : EOF
+    ch = (i < code.length()) ? code[i++].cell() : EOF
 
     QString code = in;
     QString out;
